+++ lib/netplay/netlog.h
@@ -27,6 +27,6 @@
 bool NETstartLogging();
 bool NETstopLogging();
 WZ_DECL_NONNULL(1) bool NETlogEntry(const char *str, UDWORD a, UDWORD b);
-void NETlogPacket(uint8_t type, uint32_t size, bool received);
+void NETlogPacket(uint8_t type, size_t size, bool received);
 
 #endif // _netlog_h
